+++ app/src/main/java/com/alibaba/sdk/android/oss/app/Config.java
@@ -25,4 +25,24 @@ public class Config {
     public static final String bucket = "测试bucket名称";
     public static final String uploadObject = "上传object名称";
     public static final String downloadObject = "下载object名称";
+
+    public static final int DOWNLOAD_SUC = 1;
+    public static final int DOWNLOAD_Fail = 2;
+    public static final int UPLOAD_SUC = 3;
+    public static final int UPLOAD_Fail = 4;
+    public static final int UPLOAD_PROGRESS = 5;
+    public static final int LIST_SUC = 6;
+    public static final int HEAD_SUC = 7;
+    public static final int RESUMABLE_SUC = 8;
+    public static final int SIGN_SUC = 9;
+    public static final int BUCKET_SUC = 10;
+    public static final int GET_STS_SUC = 11;
+    public static final int MULTIPART_SUC = 12;
+    public static final int STS_TOKEN_SUC = 13;
+    public static final int FAIL = 9999;
+    public static final int REQUESTCODE_AUTH = 10111;
+    public static final int REQUESTCODE_LOCALPHOTOS = 10112;
+
+
+    public static final int MESSAGE_UPLOAD_2_OSS = 10002;
 }
