+++ app/src/main/java/com/alibaba/sdk/android/oss/sample/ListObjectsSamples.java
@@ -4,12 +4,12 @@ import android.os.Handler;
 import com.alibaba.sdk.android.oss.ClientException;
 import com.alibaba.sdk.android.oss.OSS;
 import com.alibaba.sdk.android.oss.ServiceException;
-import com.alibaba.sdk.android.oss.app.MainActivity;
 import com.alibaba.sdk.android.oss.callback.OSSCompletedCallback;
 import com.alibaba.sdk.android.oss.common.OSSLog;
 import com.alibaba.sdk.android.oss.internal.OSSAsyncTask;
 import com.alibaba.sdk.android.oss.model.ListObjectsRequest;
 import com.alibaba.sdk.android.oss.model.ListObjectsResult;
+import com.alibaba.sdk.android.oss.app.Config;
 
 import java.lang.ref.WeakReference;
 
@@ -109,7 +109,7 @@ public class ListObjectsSamples extends BaseSamples{
                             + result.getObjectSummaries().get(i).getETag() + " "
                             + result.getObjectSummaries().get(i).getLastModified());
                 }
-                handler.get().sendEmptyMessage(MainActivity.LIST_SUC);
+                handler.get().sendEmptyMessage(Config.LIST_SUC);
             }
 
             @Override
@@ -126,7 +126,7 @@ public class ListObjectsSamples extends BaseSamples{
                     OSSLog.logError("HostId", serviceException.getHostId());
                     OSSLog.logError("RawMessage", serviceException.getRawMessage());
                 }
-                handler.get().sendEmptyMessage(MainActivity.FAIL);
+                handler.get().sendEmptyMessage(Config.FAIL);
             }
         });
     }
