+++ app/src/main/java/com/alibaba/sdk/android/oss/sample/ManageBucketSamples.java
@@ -4,7 +4,7 @@ import android.os.Handler;
 import com.alibaba.sdk.android.oss.ClientException;
 import com.alibaba.sdk.android.oss.OSS;
 import com.alibaba.sdk.android.oss.ServiceException;
-import com.alibaba.sdk.android.oss.app.MainActivity;
+import com.alibaba.sdk.android.oss.app.Config;
 import com.alibaba.sdk.android.oss.callback.OSSCompletedCallback;
 import com.alibaba.sdk.android.oss.common.OSSLog;
 import com.alibaba.sdk.android.oss.internal.OSSAsyncTask;
@@ -150,7 +150,7 @@ public class ManageBucketSamples extends BaseSamples{
             @Override
             public void onSuccess(DeleteBucketRequest request, DeleteBucketResult result) {
                 OSSLog.logDebug("DeleteBucket", "Success!");
-                handler.get().sendEmptyMessage(MainActivity.BUCKET_SUC);
+                handler.get().sendEmptyMessage(Config.BUCKET_SUC);
             }
 
             @Override
@@ -159,7 +159,7 @@ public class ManageBucketSamples extends BaseSamples{
                 if (clientException != null) {
                     // client side exception,  such as network exception
                     clientException.printStackTrace();
-                    handler.get().sendEmptyMessage(MainActivity.FAIL);
+                    handler.get().sendEmptyMessage(Config.FAIL);
                 }
                 if (serviceException != null) {
                     // The bucket to delete is not empty.
@@ -179,15 +179,15 @@ public class ManageBucketSamples extends BaseSamples{
                             oss.deleteBucket(deleteBucketRequest1);
                         } catch (ClientException clientexception) {
                             clientexception.printStackTrace();
-                            handler.get().sendEmptyMessage(MainActivity.FAIL);
+                            handler.get().sendEmptyMessage(Config.FAIL);
                             return;
                         } catch (ServiceException serviceexception) {
                             serviceexception.printStackTrace();
-                            handler.get().sendEmptyMessage(MainActivity.FAIL);
+                            handler.get().sendEmptyMessage(Config.FAIL);
                             return;
                         }
                         OSSLog.logDebug("DeleteBucket", "Success!");
-                        handler.get().sendEmptyMessage(MainActivity.BUCKET_SUC);
+                        handler.get().sendEmptyMessage(Config.BUCKET_SUC);
                     }
                 }
             }
